Removed the DIST_SUBDIR statement.
Explanation from maintainer:

"Log4j has been the first port from the Jakarta series and
back then I thought it would be useful to put all distfiles
from the Jakarta project in the same subdirectory. Now I
don't think it is such a good idea. Furthermore, the ports
of other Jakarta projects do not use this scheme.

So here is a patch to fix it. I didn't bump the
PORT_REVISION as there is no point in updating the port if
already installed."

PR:		56490
Submitted by:	maintainer