Update to 0.22.
This update is a collaboration between the maintainer and myself, the
libdesklets part isn't finished and needs a little more debugging (RAM
and boottime display).
Unfortunately we also have a problem with threads (they get created, but
they don't run/start) which we can't track down, so it's not easy to
debug the remaining libdesklets bugs.

We decided to commit the port in the current incarnation as at least the
desklets which I will commit shortly after this update will work without
problems.

Submitted by:	maintainer