Add nsca 2.4, Nagios Service Check Acceptor.
The Nagios Service Check Acceptor (NSCA) is used to send service check
results to a central Nagios server. This consists of the "nsca" daemon
which runs on the main Nagios server and accepts results and the
"check_nsca" plugin which is used to send results to the server.

Author:	Ethan Galstad
WWW:	http://www.nagios.org/

PR:		59436
Submitted by:	Paul Dlug <paul@nerdlabs.com>