Add CHANGES
Heavily inspired by /usr/src/UPDATING, this file is intended
for notes on major changes to ports and infrastructure, that
have a significant impact on other ports. Indended audience
is committers and developers; changes that affect users
should be added to the ports section of the release notes that
I hope to add in the near future. Although I am mentioned as
maintainer, this should not be seen as a lock, everyone is
welcome to add to this file.

Requested by:           many
Approved by:            marcus (some time ago)
Inspired by:            UPDATING (imp)