Mark as deprecated: either fails to fetch, and has done so for some time; and/or fails to build; and/or the maintainer has recommended that the ports are obsolete and should be deleted.
In each of these cases, the port has been in trouble for quite some time.
The deprecation date is set for April 09, 2004, except for a very few
ports that are in really bad shape (or the maintainer has recommended
deletion), in which case they are set for March 09, 2004.

If anyone wants to keep these ports in the ports collection, now is
the time to step up and fix them.

PR:		ports/62575
No objection:	freebsd-ports, 10 days