Update to 0.3.2.
Note: there is a sneaking suspicion abounding that this program
will not work with libc_r. If you are forcing libc_r for something
like the NVidia kernel module, update that port and switch over
to libpthread.

Note also: a noticible pause when starting a new game seems to be
normal. This pause is also associated with a gigantic memory spike,
but it'll subside once the game starts.