System_Command is a command line execution interface.
Running functions from the command line can be risky if the proper precautions
are not taken to escape the shell arguments and reaping the exit status
properly. This class provides a formal interface to both, so that you can run a
system command as comfortably as you would run a php function, with full pear
error handling as results on failure. It is important to note that this class,
unlike other implementations, distinguishes between output to stderr and output
to stdout. It also reports the exit status of the command. So in every sense of
the word, it gives php shell capabilities.

PR:		ports/76749
Submitted by:	Antonio Carlos Venancio Junior <antonio@php.net>