Update to the 20050513 snapshot of GCC 3.4.4.
Clean up MASTER_SITES.  Add conflicts with the gcc40 and gcc41 ports.
Simplify the handling of libraries which are not installed in all
configurations and put the libgcj libraries in the same directory
as all others, getting rid of the libtool .la archives on the way.
No longer install fsf-funding.7 gfdl.7 gpl.7, remove some cruft from
the post-install target, and simplify generation of the dynamic parts
of the packaging list.