Add p5-Data-Password.
Data::Password checks potential passwords for crackability. It checks that the
password has the appropriate length, that it has enough character groups, that
it does not contain the same characters repeatedly, that it does not contain
ascending or descending characters, or characters close to each other on the
keyboard. It also searches for the password in the ispell word file.

WWW: http://search.cpan.org/~razinf/Data-Password/

PR:		ports/84513
Submitted by:	ports@c0decafe.net