Add xmlprpr.
An XML pretty printer created to format XML that doesn't make use of
mixed content. In the default mode each element is put on a separate
line with consistent indentation. It can also separate attributes onto
individual lines, sort attributes in a specified or alphabetic order,
expand self closing tags, and more.

Note that the distribution calls this tool "xmlpp", but it has been
renamed so as not to conflict with an xmlpp already in the ports tree.

WWW: http://software.decisionsoft.com/tools.html

PR:		ports/92946
Submitted by:	Paul Chvostek <paul+ports@it.ca>