Add p5-Class-MOP 0.24, a Meta Object Protocol for Perl 5.
This module is an attempt to create a meta object protocol for the
Perl 5 object system. It makes no attempt to change the behavior or
characteristics of the Perl 5 object system, only to create a protocol
for its manipulation and introspection.

http://search.cpan.org/dist/Class-MOP/

PR:		ports/95740
Submitted by:	Lars Balker Rasmussen <lars@balker.dk>
Committed from:	Sofia, Bulgaria