games/gcompris2 -> games/gcompris
Rename this ports to use the real vendor package name. The advantage of this
is to allow our users' keyword search works and easier for users to file the
Bugzilla report when they use our name of ports. Debian, Gentoo, NetBSD and
other OSs have the correct package name, but not in our ports tree.

My team, FreeBSD GNOME Team, have agreed with it.