- Remove false positives for 'makefile' (and possibly a few others).
 - Remove obsolete explanations which are no longer seen, for speed:
   ELF, MOTIF, MOTIFLIB, X_manpage, awk, bison, ffs_conflict, forbidden,
   getopt, getopt.h, imake, lc_r, malloc.h, pod2man, sed, stl, soundcard.h,
   texinfo, union_wait, values.h

 - Add more cases to: arch, bad_c++, compiler_error, depend_object,
   install_error, linker_error, mtree, perl5

These changes reduce many dozens of false positives; add a few dozen
true positives; and for certain directories, improve the speed about 10%
(a few drop by 15%).

It turns out that the performance issues are mainly due to the multiple
greps.  If performance is an issue we need to go back to the moderately-
unreadable, everything-on-one-line paradigm.  Before that happens, I would
like to experiment with some refactoring, so that the patterns are built up
in the shell line-by-line, so you could still be able to read it.

Tested on:	pointyhat
Hat:		portmgr