Upgrade tclX from 8.3.5 to 8.4 released last November after several years of idleness. The changes are fairly substantial -- all Tk bits, however insignificant they were, are now removed completely, for example. No TclX enabled executable is built/installed either -- a Tcl script is expected to load the Tclx package via the "package require Tclx" only.
The changes to port add the handling of SIGINFO (made possible by the
recent fix-up of lang/tcl84), running the authors' test-target as part
of the build, and skipping building/installing of help files in the
NOPORTDOCS case.

The devel/tcl-neo is the only dependant port of tclX and is updated to
depend on the new version.